added logic so only the creator can edit/delete comments

+++ BookWormz.Services/CommentService.cs
@@ -94,6 +94,10 @@ namespace BookWormz.Services
             {
                 var entity = ctx.Comments.Single(e => e.Id == id);
 
+                //Make sure only commenter can update comment
+                if (entity.CommenterId != _userId)
+                    return false;
+
                 entity.Text = comment.CommentText;
 
                 return ctx.SaveChanges() == 1;
@@ -106,6 +110,11 @@ namespace BookWormz.Services
             using (var ctx = new ApplicationDbContext())
             {
                 var entity = ctx.Comments.Single(e => e.Id == Id);
+
+                //make sure only commenter can delete comment
+                if (entity.CommenterId != _userId)
+                    return false;
+
                 ctx.Comments.Remove(entity);
 
                 return ctx.SaveChanges() == 1;
@@ -124,7 +133,7 @@ namespace BookWormz.Services
             foreach(var reply in replies)
             {
                 var DetailedReply = new ReplyDetail { Id = reply.Id, Text = reply.Text,
-                    //Using ternary incase of comment not having author(corrupt data)
+                    //Using ternary incase of comment not having author
                     CommentorsName = (reply.Comment != null ? reply.Commenter.FullName : "Unknown") };
                 DetailedReply.Replies = AddReplies(reply.Replies);
                 DetailedReplies.Add(DetailedReply);
